Asthma in St. Louis by Census Tract, 2017

A simple features data set containing the geometry and asthma estimates from the Centers for Disease Control for St. Louis.

data(ar_stl_asthma)

Format

A data frame with 106 rows and 24 variables:

GEOID

full GEOID string

STATEFP

state FIPS code

COUNTYFP

county FIPS code

TRACTCE

tract FIPS code

NAMELSAD

tract name

ALAND

area of tract land, square meters

AWATER

area of tract water, square meters

ASTHMA

percent of residents with current asthma diagnosis, estimated

geometry

simple features geometry

Source

Centers for Disease Control's 500 Cities Data

Examples

str(ar_stl_asthma)
#> Classes ‘sf’ and 'data.frame':	106 obs. of  9 variables:
#>  $ GEOID   : chr  "29510112100" "29510116500" "29510110300" "29510103700" ...
#>  $ STATEFP : chr  "29" "29" "29" "29" ...
#>  $ COUNTYFP: chr  "510" "510" "510" "510" ...
#>  $ TRACTCE : chr  "112100" "116500" "110300" "103700" ...
#>  $ NAMELSAD: chr  "Census Tract 1121" "Census Tract 1165" "Census Tract 1103" "Census Tract 1037" ...
#>  $ ALAND   : num  6936664 904024 938287 910953 1640334 ...
#>  $ AWATER  : num  0 0 0 0 35369 ...
#>  $ ASTHMA  : num  8.8 10.8 13 8.6 9.1 8.8 14.6 12.3 10.1 13.1 ...
#>  $ geometry:sfc_POLYGON of length 106; first list element: List of 1
#>   ..$ : num [1:284, 1:2] 734633 734636 734644 734662 734671 ...
#>   ..- attr(*, "class")= chr [1:3] "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA
#>   ..- attr(*, "names")= chr [1:8] "GEOID" "STATEFP" "COUNTYFP" "TRACTCE" ...
head(ar_stl_asthma)
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> Simple feature collection with 6 features and 8 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 733360 ymin: 4274925 xmax: 741440.9 ymax: 4283867
#> Projected CRS: NAD83 / UTM zone 15N
#>         GEOID STATEFP COUNTYFP TRACTCE          NAMELSAD   ALAND AWATER ASTHMA
#> 1 29510112100      29      510  112100 Census Tract 1121 6936664      0    8.8
#> 2 29510116500      29      510  116500 Census Tract 1165  904024      0   10.8
#> 3 29510110300      29      510  110300 Census Tract 1103  938287      0   13.0
#> 4 29510103700      29      510  103700 Census Tract 1037  910953      0    8.6
#> 5 29510103800      29      510  103800 Census Tract 1038 1640334  35369    9.1
#> 6 29510104500      29      510  104500 Census Tract 1045 1939712      0    8.8
#>                         geometry
#> 1 POLYGON ((734633.3 4279479,...
#> 2 POLYGON ((740098.5 4275728,...
#> 3 POLYGON ((740110.8 4283146,...
#> 4 POLYGON ((735225.3 4276168,...
#> 5 POLYGON ((733360 4275146, 7...
#> 6 POLYGON ((735553.7 4278190,...
summary(ar_stl_asthma$ASTHMA)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    8.50    9.50   12.05   11.61   13.47   16.50