Asthma in St. Louis by Census Tract, 2017

A simple features data set containing the geometry and asthma estimates from the Centers for Disease Control for St. Louis.

data(ar_stl_asthma)

Format

A data frame with 106 rows and 24 variables:

GEOID

full GEOID string

STATEFP

state FIPS code

COUNTYFP

county FIPS code

TRACTCE

tract FIPS code

NAMELSAD

tract name

ALAND

area of tract land, square meters

AWATER

area of tract water, square meters

ASTHMA

percent of residents with current asthma diagnosis, estimated

geometry

simple features geometry

Source

Centers for Disease Control's 500 Cities Data

Examples

str(ar_stl_asthma)
#> Classes ‘sf’ and 'data.frame':	106 obs. of  9 variables:
#>  $ GEOID   : chr  "29510102300" "29510102400" "29510104500" "29510106100" ...
#>  $ STATEFP : chr  "29" "29" "29" "29" ...
#>  $ COUNTYFP: chr  "510" "510" "510" "510" ...
#>  $ TRACTCE : chr  "102300" "102400" "104500" "106100" ...
#>  $ NAMELSAD: chr  "Census Tract 1023" "Census Tract 1024" "Census Tract 1045" "Census Tract 1061" ...
#>  $ ALAND   : num  1282334 584695 1939714 962329 1195496 ...
#>  $ AWATER  : num  49882 0 0 0 0 ...
#>  $ ASTHMA  : num  9.2 10.3 8.8 14.6 12.3 10.1 13.1 12.5 8.6 9.6 ...
#>  $ geometry:sfc_POLYGON of length 106; first list element: List of 1
#>   ..$ : num [1:137, 1:2] 267690 267696 267704 267727 267735 ...
#>   ..- attr(*, "class")= chr [1:3] "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA
#>   ..- attr(*, "names")= chr [1:8] "GEOID" "STATEFP" "COUNTYFP" "TRACTCE" ...
head(ar_stl_asthma)
#> Simple feature collection with 6 features and 8 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 267690.4 ymin: 302101.6 xmax: 270505.9 ymax: 315447
#> Projected CRS: NAD_1983_HARN_StatePlane_Missouri_East_FIPS_2401
#>         GEOID STATEFP COUNTYFP TRACTCE          NAMELSAD   ALAND AWATER ASTHMA
#> 1 29510102300      29      510  102300 Census Tract 1023 1282334  49882    9.2
#> 2 29510102400      29      510  102400 Census Tract 1024  584695      0   10.3
#> 3 29510104500      29      510  104500 Census Tract 1045 1939714      0    8.8
#> 4 29510106100      29      510  106100 Census Tract 1061  962329      0   14.6
#> 5 29510105500      29      510  105500 Census Tract 1055 1195496      0   12.3
#> 6 29510105200      29      510  105200 Census Tract 1052  830997      0   10.1
#>                         geometry
#> 1 POLYGON ((267690.4 302870.7...
#> 2 POLYGON ((268776.8 304170, ...
#> 3 POLYGON ((267911.2 309377, ...
#> 4 POLYGON ((268270.7 314883.8...
#> 5 POLYGON ((268625.5 313598, ...
#> 6 POLYGON ((267878.6 312440.7...
summary(ar_stl_asthma$ASTHMA)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    8.50    9.50   12.05   11.61   13.47   16.50