Race in St. Louis by Census Tract, 2017

A simple features data set containing the geometry and associated attributes for the 2013-2017 American Community Survey estimates for race in St. Louis.

data(ar_stl_race)

Format

A data frame with 106 rows and 24 variables:

GEOID

full GEOID string

STATEFP

state FIPS code

COUNTYFP

county FIPS code

TRACTCE

tract FIPS code

NAMELSAD

tract name

ALAND

area of tract land, square meters

AWATER

area of tract water, square meters

TOTAL_E

total populaton count, estimated

TOTAL_M

total populaton count, margin of error

WHITE_E

white populaton count, estimated

WHITE_M

white populaton count, margin of error

BLACK_E

black populaton count, estimated

BLACK_M

black populaton count, margin of error

AIAN_E

american indian and alskan native populaton count, estimated

AIAN_M

american indian and alskan native populaton count, margin of error

ASIAN_E

asian populaton count, estimated

ASIAN_M

asian populaton count, margin of error

NHPI_E

native hawaiian and pacific islander populaton count, estimated

NHPI_M

native hawaiian and pacific islander populaton count, margin of error

OTHER_E

other race populaton count, estimated

OTHER_M

other race populaton count, margin of error

TWOPLUS_E

two or more races populaton count, estimated

TWOPLUS_M

two or more races populaton count, margin of error

geometry

simple features geometry

Source

tidycensus package

Examples

str(ar_stl_race)
#> Classes ‘sf’ and 'data.frame':	106 obs. of  24 variables:
#>  $ GEOID    : chr  "29510102300" "29510102400" "29510104500" "29510106100" ...
#>  $ STATEFP  : chr  "29" "29" "29" "29" ...
#>  $ COUNTYFP : chr  "510" "510" "510" "510" ...
#>  $ TRACTCE  : chr  "102300" "102400" "104500" "106100" ...
#>  $ NAMELSAD : chr  "Census Tract 1023" "Census Tract 1024" "Census Tract 1045" "Census Tract 1061" ...
#>  $ ALAND    : num  1282334 584695 1939714 962329 1195496 ...
#>  $ AWATER   : num  49882 0 0 0 0 ...
#>  $ TOTAL_E  : num  2126 2499 1990 1826 2819 ...
#>  $ TOTAL_M  : num  207 180 225 285 488 308 408 279 227 166 ...
#>  $ WHITE_E  : num  1763 2252 1487 32 67 ...
#>  $ WHITE_M  : num  227 227 195 25 56 215 198 233 229 144 ...
#>  $ BLACK_E  : num  97 217 162 1787 2713 ...
#>  $ BLACK_M  : num  83 167 125 281 493 221 388 251 86 30 ...
#>  $ AIAN_E   : num  0 0 1 0 0 39 0 0 9 0 ...
#>  $ AIAN_M   : num  11 11 3 11 11 60 11 11 14 11 ...
#>  $ ASIAN_E  : num  237 0 280 0 24 196 40 0 217 12 ...
#>  $ ASIAN_M  : num  143 11 132 11 28 98 36 11 105 13 ...
#>  $ NHPI_E   : num  0 0 0 7 0 0 0 5 16 0 ...
#>  $ NHPI_M   : num  11 11 11 13 11 11 11 7 23 11 ...
#>  $ OTHER_E  : num  16 0 0 0 0 0 0 90 31 11 ...
#>  $ OTHER_M  : num  25 11 11 11 11 11 11 72 36 12 ...
#>  $ TWOPLUS_E: num  13 30 60 0 15 120 66 122 55 49 ...
#>  $ TWOPLUS_M: num  22 49 39 11 33 97 96 86 50 68 ...
#>  $ geometry :sfc_POLYGON of length 106; first list element: List of 1
#>   ..$ : num [1:137, 1:2] 735510 735516 735523 735546 735554 ...
#>   ..- attr(*, "class")= chr [1:3] "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr [1:23] "GEOID" "STATEFP" "COUNTYFP" "TRACTCE" ...
head(ar_stl_race)
#> Simple feature collection with 6 features and 23 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 735419.3 ymin: 4270933 xmax: 738124.7 ymax: 4284280
#> Projected CRS: NAD83 / UTM zone 15N
#>         GEOID STATEFP COUNTYFP TRACTCE          NAMELSAD   ALAND AWATER TOTAL_E
#> 1 29510102300      29      510  102300 Census Tract 1023 1282334  49882    2126
#> 2 29510102400      29      510  102400 Census Tract 1024  584695      0    2499
#> 3 29510104500      29      510  104500 Census Tract 1045 1939714      0    1990
#> 4 29510106100      29      510  106100 Census Tract 1061  962329      0    1826
#> 5 29510105500      29      510  105500 Census Tract 1055 1195496      0    2819
#> 6 29510105200      29      510  105200 Census Tract 1052  830997      0    2981
#>   TOTAL_M WHITE_E WHITE_M BLACK_E BLACK_M AIAN_E AIAN_M ASIAN_E ASIAN_M NHPI_E
#> 1     207    1763     227      97      83      0     11     237     143      0
#> 2     180    2252     227     217     167      0     11       0      11      0
#> 3     225    1487     195     162     125      1      3     280     132      0
#> 4     285      32      25    1787     281      0     11       0      11      7
#> 5     488      67      56    2713     493      0     11      24      28      0
#> 6     308    1098     215    1528     221     39     60     196      98      0
#>   NHPI_M OTHER_E OTHER_M TWOPLUS_E TWOPLUS_M                       geometry
#> 1     11      16      25        13        22 POLYGON ((735510.2 4271678,...
#> 2     11       0      11        30        49 POLYGON ((736561.2 4273007,...
#> 3     11       0      11        60        39 POLYGON ((735553.7 4278190,...
#> 4     13       0      11         0        11 POLYGON ((735763.1 4283707,...
#> 5     11       0      11        15        33 POLYGON ((736153 4282431, 7...
#> 6     11       0      11       120        97 POLYGON ((735437.7 4281253,...
summary(ar_stl_race$ALAND)
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#>   506376   913345  1121825  1508633  1651801 10277994