Race in St. Louis by Census Tract, 2017

A simple features data set containing the geometry and associated attributes for the 2013-2017 American Community Survey estimates for race in St. Louis.

data(ar_stl_race)

Format

A data frame with 106 rows and 24 variables:

GEOID

full GEOID string

STATEFP

state FIPS code

COUNTYFP

county FIPS code

TRACTCE

tract FIPS code

NAMELSAD

tract name

ALAND

area of tract land, square meters

AWATER

area of tract water, square meters

TOTAL_E

total populaton count, estimated

TOTAL_M

total populaton count, margin of error

WHITE_E

white populaton count, estimated

WHITE_M

white populaton count, margin of error

BLACK_E

black populaton count, estimated

BLACK_M

black populaton count, margin of error

AIAN_E

american indian and alskan native populaton count, estimated

AIAN_M

american indian and alskan native populaton count, margin of error

ASIAN_E

asian populaton count, estimated

ASIAN_M

asian populaton count, margin of error

NHPI_E

native hawaiian and pacific islander populaton count, estimated

NHPI_M

native hawaiian and pacific islander populaton count, margin of error

OTHER_E

other race populaton count, estimated

OTHER_M

other race populaton count, margin of error

TWOPLUS_E

two or more races populaton count, estimated

TWOPLUS_M

two or more races populaton count, margin of error

geometry

simple features geometry

Source

tidycensus package

Examples

str(ar_stl_race)

#> Classes ‘sf’ and 'data.frame':	106 obs. of  24 variables:
#>  $ GEOID    : chr  "29510112100" "29510116500" "29510110300" "29510103700" ...
#>  $ STATEFP  : chr  "29" "29" "29" "29" ...
#>  $ COUNTYFP : chr  "510" "510" "510" "510" ...
#>  $ TRACTCE  : chr  "112100" "116500" "110300" "103700" ...
#>  $ NAMELSAD : chr  "Census Tract 1121" "Census Tract 1165" "Census Tract 1103" "Census Tract 1037" ...
#>  $ ALAND    : num  6936664 904024 938287 910953 1640334 ...
#>  $ AWATER   : num  0 0 0 0 35369 ...
#>  $ TOTAL_E  : num  4253 4590 1936 2406 3856 ...
#>  $ TOTAL_M  : num  468 525 360 215 242 225 285 488 308 408 ...
#>  $ WHITE_E  : num  2827 2194 47 2184 3383 ...
#>  $ WHITE_M  : num  378 411 40 255 269 195 25 56 215 198 ...
#>  $ BLACK_E  : num  1147 1885 1889 122 300 ...
#>  $ BLACK_M  : num  393 414 361 117 208 125 281 493 221 388 ...
#>  $ AIAN_E   : num  0 0 0 0 0 1 0 0 39 0 ...
#>  $ AIAN_M   : num  11 11 11 11 11 3 11 11 60 11 ...
#>  $ ASIAN_E  : num  179 279 0 48 62 280 0 24 196 40 ...
#>  $ ASIAN_M  : num  86 172 11 55 60 132 11 28 98 36 ...
#>  $ NHPI_E   : num  0 0 0 0 0 0 7 0 0 0 ...
#>  $ NHPI_M   : num  11 11 11 11 11 11 13 11 11 11 ...
#>  $ OTHER_E  : num  15 59 0 0 77 0 0 0 0 0 ...
#>  $ OTHER_M  : num  23 67 11 11 120 11 11 11 11 11 ...
#>  $ TWOPLUS_E: num  85 173 0 52 34 60 0 15 120 66 ...
#>  $ TWOPLUS_M: num  68 126 11 49 38 39 11 33 97 96 ...
#>  $ geometry :sfc_POLYGON of length 106; first list element: List of 1
#>   ..$ : num [1:284, 1:2] 734633 734636 734644 734662 734671 ...
#>   ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr  "GEOID" "STATEFP" "COUNTYFP" "TRACTCE" ...

head(ar_stl_race)

#> Simple feature collection with 6 features and 23 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 733360 ymin: 4274925 xmax: 741440.9 ymax: 4283867
#> epsg (SRID):    26915
#> proj4string:    +proj=utm +zone=15 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#>         GEOID STATEFP COUNTYFP TRACTCE          NAMELSAD   ALAND AWATER TOTAL_E
#> 1 29510112100      29      510  112100 Census Tract 1121 6936664      0    4253
#> 2 29510116500      29      510  116500 Census Tract 1165  904024      0    4590
#> 3 29510110300      29      510  110300 Census Tract 1103  938287      0    1936
#> 4 29510103700      29      510  103700 Census Tract 1037  910953      0    2406
#> 5 29510103800      29      510  103800 Census Tract 1038 1640334  35369    3856
#> 6 29510104500      29      510  104500 Census Tract 1045 1939712      0    1990
#>   TOTAL_M WHITE_E WHITE_M BLACK_E BLACK_M AIAN_E AIAN_M ASIAN_E ASIAN_M NHPI_E
#> 1     468    2827     378    1147     393      0     11     179      86      0
#> 2     525    2194     411    1885     414      0     11     279     172      0
#> 3     360      47      40    1889     361      0     11       0      11      0
#> 4     215    2184     255     122     117      0     11      48      55      0
#> 5     242    3383     269     300     208      0     11      62      60      0
#> 6     225    1487     195     162     125      1      3     280     132      0
#>   NHPI_M OTHER_E OTHER_M TWOPLUS_E TWOPLUS_M                       geometry
#> 1     11      15      23        85        68 POLYGON ((734633.3 4279479,...
#> 2     11      59      67       173       126 POLYGON ((740098.5 4275728,...
#> 3     11       0      11         0        11 POLYGON ((740110.8 4283146,...
#> 4     11       0      11        52        49 POLYGON ((735225.3 4276168,...
#> 5     11      77     120        34        38 POLYGON ((733360 4275146, 7...
#> 6     11       0      11        60        39 POLYGON ((735553.7 4278190,...

summary(ar_stl_race$ALAND)

#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#>   506383   913320  1121494  1513755  1651019 10530448