Clipped Ward Boundaries in St. Louis, 2010

A simple features data set containing the 2010 Ward boundaries, which are used as districts for Alderpersons who serve as elected representatives. This version of the ward boundary has been modified so that the wards only extend to the Mississippi River shoreline.

data(ar_stl_wardsClipped)

Format

A data frame with 28 rows and 2 variables:

WARD

Ward number

geometry

simple features geometry

Source

City of St. Louis

Examples

str(ar_stl_wardsClipped)

#> Classes ‘sf’ and 'data.frame':	28 obs. of  2 variables:
#>  $ WARD    : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ geometry:sfc_POLYGON of length 28; first list element: List of 1
#>   ..$ : num [1:171, 1:2] 740184 740220 740295 740375 740541 ...
#>   ..- attr(*, "class")= chr  "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   ..- attr(*, "names")= chr "WARD"

head(ar_stl_wardsClipped)

#> Simple feature collection with 6 features and 1 field
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 737723.1 ymin: 4275825 xmax: 745442.5 ymax: 4295504
#> epsg (SRID):    26915
#> proj4string:    +proj=utm +zone=15 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#>   WARD                       geometry
#> 1    1 POLYGON ((740184.2 4286431,...
#> 2    2 POLYGON ((745382.7 4280783,...
#> 3    3 POLYGON ((742956.1 4284113,...
#> 4    4 POLYGON ((739557.6 4284080,...
#> 5    5 POLYGON ((744883.8 4281632,...
#> 6    6 POLYGON ((742501.6 4279976,...