Clipped Ward Boundaries in St. Louis, 2010

A simple features data set containing the 2010 Ward boundaries, which are used as districts for Alderpersons who serve as elected representatives. This version of the ward boundary has been modified so that the wards only extend to the Mississippi River shoreline.

data(ar_stl_wardsClipped)

Format

A data frame with 28 rows and 2 variables:

WARD

Ward number

geometry

simple features geometry

Source

City of St. Louis

Examples

str(ar_stl_wardsClipped)
#> Classes ‘sf’ and 'data.frame':	28 obs. of  2 variables:
#>  $ WARD    : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ geometry:sfc_POLYGON of length 28; first list element: List of 1
#>   ..$ : num [1:171, 1:2] 272762 272796 272868 272946 273105 ...
#>   ..- attr(*, "class")= chr [1:3] "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   ..- attr(*, "names")= chr "WARD"
head(ar_stl_wardsClipped)
#> Simple feature collection with 6 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 270215.8 ymin: 306873.8 xmax: 278218 ymax: 326431.1
#> Projected CRS: NAD_1983_HARN_StatePlane_Missouri_East_FIPS_2401
#>   WARD                       geometry
#> 1    1 POLYGON ((272762.3 317485.6...
#> 2    2 POLYGON ((277802.9 311701.1...
#> 3    3 POLYGON ((275468.9 315094.7...
#> 4    4 POLYGON ((272072.1 315154, ...
#> 5    5 POLYGON ((277327.5 312563, ...
#> 6    6 POLYGON ((274902 310972.4, ...